fix: Setting the ROI after the exposure value renders the auto exposure non-effective.
This fixes a bug that is introduced by setting the region of interest after
the exposure mode and values are set. For the IDS cameras, setting exposure
values will deactivate the auto exposure mode if set previously.
In the super class method 'configureCamera' the exposure values are correctly
set after the ROI is applied. Reusing the super class method therefore fixes
this bug and prevents duplication of code.